--- module/core/former/tests/inc/enum_tuple_zero_field_tests/mod.rs
@@ -1,3 +1,0 @@
-//! Tests for zero-field tuple variants, e.g., `MyEnum::Variant()`.
-mod tuple_zero_fields_derive;
-mod tuple_zero_fields_manual;
