+++ module/core/former/tests/inc/enum_unnamed_tests/mod.rs
@@ -84,8 +84,12 @@
 // mod standalone_constructor_args_tuple_single_manual; // Added
 // mod standalone_constructor_args_tuple_multi_manual; // Added
 // mod standalone_constructor_args_tuple_only_test;
-// mod tuple_zero_fields_derive; // Moved from enum_unit_tests
-// mod tuple_zero_fields_manual; // Moved from enum_unit_tests
-// mod tuple_zero_fields_only_test; // Moved from enum_unit_tests
+
+// Coverage for `tuple_zero_fields_*` tests:
+// - Tests zero-field tuple variants e.g., `MyEnum::Variant()`.
+// - Verifies Rules 1b (scalar), 3b (default), and 4a (standalone_constructors).
+// mod tuple_zero_fields_derive; // Moved from enum_unit_tests - Temporarily commented for debugging Access Violation
+// mod tuple_zero_fields_manual; // Moved from enum_unit_tests - Temporarily commented for debugging Access Violation
+// Note: tuple_zero_fields_only_test.rs is included by the manual and derive files.
 
 // pub mod compile_fail;
